add:  multiplyGpu implemented

+++ TakiMatrix/includes/kernel/kernel.h
@@ -8,12 +8,12 @@
 #include <cuda_runtime.h>
 #include <device_launch_parameters.h>
 
-__global__ void addKernel(const float* first, const float* second, unsigned int size);
+__global__ void addKernel(const float* first, const float* second, size_t size);
 
-__global__ void subKernel(const float* first, const float* second, unsigned int size);
+__global__ void subKernel(const float* first, const float* second, size_t size);
 
 __global__ void multiplyKernel(const float* first, const float* second,
-        float* result, unsigned int firstRowNum,
-        unsigned int middle, unsigned int secondColNum);
+        float* result, size_t firstRowNum,
+        size_t middle, size_t secondColNum);
 
 #endif // TAKIMATRIX_KERNEL_H
